requirements.txt
requests==2.31.0
pandas==2.2.0
boto3==1.34.0
APScheduler==3.10.4
python-dotenv==1.0.0
loguru==0.7.2